# Add measurement of zipping performance