Added test case to measure zipping performance. 

## **Performance Testing** 

Performed performance test on a sample **text file containing ASCII characters**. Test results will vary for different file types and devices. Below are the test results for illustration.

**Test device** - iPhone 4s (A5 chipset, 512MB RAM, Dual-core 1GHz Cortex-A9)

| Original file size | Time | Compressed file size |
| --- | --- | --- |
| 1 MB | 0.201s | 150 KB |
| 2.5 MB | 0.476s | 360 KB |
| 5 MB | 0.948s | 744 KB | 
| 7.5 MB | 1.402s | 1.08 MB |
| 10 MB | 1.919s | 1.48 MB |
| 25 MB | 4.641s | 3.7 MB |