@Coeur 2018-04-03
Thank you, but this is not a Unit Test, it's benchmarking. So it could be somewhere else, but not in SSZipArchiveTests.m. And if it requires a 25 MB file for testing it, then we should:
* either have this file randomly generated (problem: results won't be consistent across devices)
* either move it to a separate repository: https://github.com/ZipArchive/ZipArchive/issues/399