@Coeur 2018-08-29
Closing for now, as it's not a unit test in its current state.